Fix bug xong gặp bug =)))

+++ src/main/java/JavaWeb/GameAccount/controllers/ProductController.java
@@ -43,7 +43,7 @@ public class ProductController {
     }
 
     @GetMapping("/edit/{id}")
-    public String showEditForm(@PathVariable Long id, Model model) {
+    public String showEditForm(@PathVariable int id, Model model) {
         Product product = productService.getProductById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid productId:" + id));
         model.addAttribute("product", product);
